+++ tensorflow_probability/examples/bayesian_neural_network.py
@@ -32,16 +32,16 @@ import os
 import warnings
 
 # Dependency imports
-from absl import flags
+from absl import flags, app
 import matplotlib
 matplotlib.use("Agg")
 from matplotlib import figure  # pylint: disable=g-import-not-at-top
 from matplotlib.backends import backend_agg
 import numpy as np
-import tensorflow as tf
+import tensorflow.compat.v2 as tf
 import tensorflow_probability as tfp
 
-from tensorflow.contrib.learn.python.learn.datasets import mnist
+tf.enable_v2_behavior()
 
 # TODO(b/78137893): Integration tests currently fail with seaborn imports.
 warnings.simplefilter(action="ignore")
@@ -55,12 +55,15 @@ except ImportError:
 tfd = tfp.distributions
 
 IMAGE_SHAPE = [28, 28, 1]
+NUM_TRAIN_EXAMPLES = 60000
+NUM_HELDOUT_EXAMPLES = 10000
+NUM_CLASSES = 10
 
 flags.DEFINE_float("learning_rate",
                    default=0.001,
                    help="Initial learning rate.")
-flags.DEFINE_integer("max_steps",
-                     default=6000,
+flags.DEFINE_integer("num_epochs",
+                     default=10,
                      help="Number of training steps to run.")
 flags.DEFINE_integer("batch_size",
                      default=128,
@@ -81,7 +84,7 @@ flags.DEFINE_integer("num_monte_carlo",
                      default=50,
                      help="Network draws to compute predictive probabilities.")
 flags.DEFINE_bool("fake_data",
-                  default=None,
+                  default=False,
                   help="If true, uses fake data. Defaults to real data.")
 
 FLAGS = flags.FLAGS
@@ -93,11 +96,11 @@ def plot_weight_posteriors(names, qm_vals, qs_vals, fname):
   Args:
     names: A Python `iterable` of `str` variable names.
     qm_vals: A Python `iterable`, the same length as `names`,
-      whose elements are Numpy `array`s, of any shape, containing
-      posterior means of weight varibles.
+    whose elements are Numpy `array`s, of any shape, containing
+    posterior means of weight varibles.
     qs_vals: A Python `iterable`, the same length as `names`,
-      whose elements are Numpy `array`s, of any shape, containing
-      posterior standard deviations of weight varibles.
+    whose elements are Numpy `array`s, of any shape, containing
+    posterior standard deviations of weight varibles.
     fname: Python `str` filename to save the plot to.
   """
   fig = figure.Figure(figsize=(6, 3))
@@ -105,14 +108,14 @@ def plot_weight_posteriors(names, qm_vals, qs_vals, fname):
 
   ax = fig.add_subplot(1, 2, 1)
   for n, qm in zip(names, qm_vals):
-    sns.distplot(qm.flatten(), ax=ax, label=n)
+    sns.distplot(qm.numpy().flatten(), ax=ax, label=n)
   ax.set_title("weight means")
   ax.set_xlim([-1.5, 1.5])
   ax.legend()
 
   ax = fig.add_subplot(1, 2, 2)
   for n, qs in zip(names, qs_vals):
-    sns.distplot(qs.flatten(), ax=ax)
+    sns.distplot(qs.numpy().flatten(), ax=ax)
   ax.set_title("weight stddevs")
   ax.set_xlim([0, 1.])
 
@@ -127,10 +130,10 @@ def plot_heldout_prediction(input_vals, probs,
 
   Args:
     input_vals: A `float`-like Numpy `array` of shape
-      `[num_heldout] + IMAGE_SHAPE`, containing heldout input images.
+    `[num_heldout] + IMAGE_SHAPE`, containing heldout input images.
     probs: A `float`-like Numpy array of shape `[num_monte_carlo,
-      num_heldout, num_classes]` containing Monte Carlo samples of
-      class probabilities for each heldout sample.
+    num_heldout, num_classes]` containing Monte Carlo samples of
+    class probabilities for each heldout sample.
     fname: Python `str` filename to save the plot to.
     n: Python `int` number of datapoints to vizualize.
     title: Python `str` title for the plot.
@@ -158,56 +161,131 @@ def plot_heldout_prediction(input_vals, probs,
   print("saved {}".format(fname))
 
 
-def build_input_pipeline(mnist_data, batch_size, heldout_size):
-  """Build an Iterator switching between train and heldout data."""
+def create_model():
+  """
+  Creates a Keras model using the LeNet-5 architecture.
 
-  # Build an iterator over training batches.
-  training_dataset = tf.data.Dataset.from_tensor_slices(
-      (mnist_data.train.images, np.int32(mnist_data.train.labels)))
-  training_batches = training_dataset.shuffle(
-      50000, reshuffle_each_iteration=True).repeat().batch(batch_size)
-  training_iterator = tf.compat.v1.data.make_one_shot_iterator(training_batches)
+  Returns:
+      model: Compiled Keras model.
+  """
+  # KL divergence weighted by the number of training samples, using
+  # lambda function to pass as input to the kernel_divergence_fn on
+  # flipout layers.
+  kl_divergence_function = lambda q, p, _: tfd.kl_divergence(q, p) / \
+      tf.cast(NUM_TRAIN_EXAMPLES, dtype=tf.float32)
 
-  # Build a iterator over the heldout set with batch_size=heldout_size,
-  # i.e., return the entire heldout set as a constant.
-  heldout_dataset = tf.data.Dataset.from_tensor_slices(
-      (mnist_data.validation.images,
-       np.int32(mnist_data.validation.labels)))
-  heldout_frozen = (heldout_dataset.take(heldout_size).
-                    repeat().batch(heldout_size))
-  heldout_iterator = tf.compat.v1.data.make_one_shot_iterator(heldout_frozen)
+  # Define a LeNet-5 model using three convolutional (with max pooling)
+  # and two fully connected dense layers. We use the Flipout
+  # Monte Carlo estimator for these layers, which enables lower variance
+  # stochastic gradients than naive reparameterization.
+  model = tf.keras.models.Sequential()
+  model.add(tfp.layers.Convolution2DFlipout(
+      6, kernel_size=5, padding="SAME",
+      kernel_divergence_fn=kl_divergence_function,
+      activation=tf.nn.relu))
+  model.add(tf.keras.layers.MaxPooling2D(
+      pool_size=[2, 2], strides=[2, 2],
+      padding="SAME"))
+  model.add(tfp.layers.Convolution2DFlipout(
+      16, kernel_size=5, padding="SAME",
+      kernel_divergence_fn=kl_divergence_function,
+      activation=tf.nn.relu))
+  model.add(tf.keras.layers.MaxPooling2D(
+      pool_size=[2, 2], strides=[2, 2],
+      padding="SAME"))
+  model.add(tfp.layers.Convolution2DFlipout(
+      120, kernel_size=5, padding="SAME",
+      kernel_divergence_fn=kl_divergence_function,
+      activation=tf.nn.relu))
+  model.add(tf.keras.layers.Flatten())
+  model.add(tfp.layers.DenseFlipout(
+      84, kernel_divergence_fn=kl_divergence_function,
+      activation=tf.nn.relu))
+  model.add(tfp.layers.DenseFlipout(
+      NUM_CLASSES, kernel_divergence_fn=kl_divergence_function,
+      activation=tf.nn.softmax))
 
-  # Combine these into a feedable iterator that can switch between training
-  # and validation inputs.
-  handle = tf.compat.v1.placeholder(tf.string, shape=[])
-  feedable_iterator = tf.compat.v1.data.Iterator.from_string_handle(
-      handle, training_batches.output_types, training_batches.output_shapes)
-  images, labels = feedable_iterator.get_next()
+  # Model compilation.
+  optimizer = tf.keras.optimizers.Adam(lr=FLAGS.learning_rate)
+  # We use the categorical_crossentropy loss since the MNIST dataset contains
+  # ten labels. The Keras API will then automatically add the
+  # Kullback-Leibler divergence (contained on the individual layers of
+  # the model), to the cross entropy loss, effectively
+  # calcuating the (negated) Evidence Lower Bound Loss (ELBO)
+  model.compile(optimizer, loss='categorical_crossentropy',
+                metrics=['accuracy'])
+  return model
 
-  return images, labels, handle, training_iterator, heldout_iterator
 
+class MNISTSequence(tf.keras.utils.Sequence):
+  """
+  Class inheriting from the Keras Sequence class for producing
+  a sequence of data (to be used with a model.(fit|evaluate|predict)_generator
+  function).
 
-def build_fake_data(num_examples=10):
-  """Build fake MNIST-style data for unit testing."""
+  Args:
+    data: Tuple of numpy `array` instances, the first representing images and
+          the second labels.
+    batch_size: Integer, number of elements in each training batch.
+    fake_data: Boolean, for creating MNIST-like fake data.
+  """
+  def __init__(self, data=None, batch_size=128, size=None):
+    if data:
+      images, labels = data
+    else:
+      images, labels = MNISTSequence.__generate_fake_data(
+          size, NUM_CLASSES)
+    self.images, self.labels = MNISTSequence.__preprocessing(
+        images, labels)
+    self.batch_size = batch_size
 
-  class Dummy(object):
-    pass
+  @staticmethod
+  def __generate_fake_data(num_images, num_classes):
+    '''
+    Generate fake data in the shape of the MNIST dataset for unittest.
+    Args:
+      num_images: Integer, the number of fake images to be generated.
+    Returns:
+      images: Numpy `array` representing the fake image data. The
+              shape of the array will be (num_images, 28, 28).
+      labels: Numpy `array` of integers, where each entry will be
+              assigned a unique integer.
+    '''
+    images = np.random.randint(low=0, high=256,
+                               size=(num_images, IMAGE_SHAPE[0],
+                                     IMAGE_SHAPE[1]))
+    labels = np.random.randint(low=0, high=num_classes,
+                               size=num_images)
+    return images, labels
 
-  num_examples = 10
-  mnist_data = Dummy()
-  mnist_data.train = Dummy()
-  mnist_data.train.images = np.float32(np.random.randn(
-      num_examples, *IMAGE_SHAPE))
-  mnist_data.train.labels = np.int32(np.random.permutation(
-      np.arange(num_examples)))
-  mnist_data.train.num_examples = num_examples
-  mnist_data.validation = Dummy()
-  mnist_data.validation.images = np.float32(np.random.randn(
-      num_examples, *IMAGE_SHAPE))
-  mnist_data.validation.labels = np.int32(np.random.permutation(
-      np.arange(num_examples)))
-  mnist_data.validation.num_examples = num_examples
-  return mnist_data
+  @staticmethod
+  def __preprocessing(images, labels):
+    '''
+    Preprocessing for image and labels data.
+
+    Args:
+      images: Numpy `array` representing the image data.
+      labels: Numpy `array` representing the labels data (range 0-9).
+
+    Returns:
+      images: Numpy `array` representing the image data, normalized
+              and expanded for convolutional network input.
+      labels: Numpy `array` representing the labels data (range 0-9),
+              as one-hot (categorical) values.
+    '''
+    images = images/255.
+    images = np.expand_dims(images, axis=4)
+
+    labels = tf.keras.utils.to_categorical(labels)
+    return images, labels
+
+  def __len__(self):
+    return int(np.ceil(len(self.images)/self.batch_size))
+
+  def __getitem__(self, idx):
+    batch_x = self.images[idx*self.batch_size:(idx+1)*self.batch_size]
+    batch_y = self.labels[idx*self.batch_size:(idx+1)*self.batch_size]
+    return batch_x, batch_y
 
 
 def main(argv):
@@ -219,125 +297,63 @@ def main(argv):
   tf.io.gfile.makedirs(FLAGS.model_dir)
 
   if FLAGS.fake_data:
-    mnist_data = build_fake_data()
+    train_seq = MNISTSequence(batch_size=FLAGS.batch_size,
+                              size=NUM_TRAIN_EXAMPLES)
+    heldout_seq = MNISTSequence(batch_size=FLAGS.batch_size,
+                                size=NUM_HELDOUT_EXAMPLES)
   else:
-    mnist_data = mnist.read_data_sets(FLAGS.data_dir, reshape=False)
+    train_set, heldout_set = tf.keras.datasets.mnist.load_data()
+    train_seq = MNISTSequence(data=train_set, batch_size=FLAGS.batch_size)
+    heldout_seq = MNISTSequence(data=heldout_set, batch_size=FLAGS.batch_size)
 
-  (images, labels, handle,
-   training_iterator, heldout_iterator) = build_input_pipeline(
-       mnist_data, FLAGS.batch_size, mnist_data.validation.num_examples)
+  model = create_model()
 
-  # Build a Bayesian LeNet5 network. We use the Flipout Monte Carlo estimator
-  # for the convolution and fully-connected layers: this enables lower
-  # variance stochastic gradients than naive reparameterization.
-  with tf.compat.v1.name_scope("bayesian_neural_net", values=[images]):
-    neural_net = tf.keras.Sequential([
-        tfp.layers.Convolution2DFlipout(6,
-                                        kernel_size=5,
-                                        padding="SAME",
-                                        activation=tf.nn.relu),
-        tf.keras.layers.MaxPooling2D(pool_size=[2, 2],
-                                     strides=[2, 2],
-                                     padding="SAME"),
-        tfp.layers.Convolution2DFlipout(16,
-                                        kernel_size=5,
-                                        padding="SAME",
-                                        activation=tf.nn.relu),
-        tf.keras.layers.MaxPooling2D(pool_size=[2, 2],
-                                     strides=[2, 2],
-                                     padding="SAME"),
-        tfp.layers.Convolution2DFlipout(120,
-                                        kernel_size=5,
-                                        padding="SAME",
-                                        activation=tf.nn.relu),
-        tf.keras.layers.Flatten(),
-        tfp.layers.DenseFlipout(84, activation=tf.nn.relu),
-        tfp.layers.DenseFlipout(10)
-        ])
-
-    logits = neural_net(images)
-    labels_distribution = tfd.Categorical(logits=logits)
-
-  # Compute the -ELBO as the loss, averaged over the batch size.
-  neg_log_likelihood = -tf.reduce_mean(
-      input_tensor=labels_distribution.log_prob(labels))
-  kl = sum(neural_net.losses) / mnist_data.train.num_examples
-  elbo_loss = neg_log_likelihood + kl
-
-  # Build metrics for evaluation. Predictions are formed from a single forward
-  # pass of the probabilistic layers. They are cheap but noisy predictions.
-  predictions = tf.argmax(input=logits, axis=1)
-  accuracy, accuracy_update_op = tf.compat.v1.metrics.accuracy(
-      labels=labels, predictions=predictions)
-
-  # Extract weight posterior statistics for layers with weight distributions
-  # for later visualization.
-  names = []
-  qmeans = []
-  qstds = []
-  for i, layer in enumerate(neural_net.layers):
-    try:
-      q = layer.kernel_posterior
-    except AttributeError:
-      continue
-    names.append("Layer {}".format(i))
-    qmeans.append(q.mean())
-    qstds.append(q.stddev())
-
-  with tf.compat.v1.name_scope("train"):
-    optimizer = tf.compat.v1.train.AdamOptimizer(
-        learning_rate=FLAGS.learning_rate)
-    train_op = optimizer.minimize(elbo_loss)
-
-  init_op = tf.group(tf.compat.v1.global_variables_initializer(),
-                     tf.compat.v1.local_variables_initializer())
-
-  with tf.compat.v1.Session() as sess:
-    sess.run(init_op)
-
-    # Run the training loop.
-    train_handle = sess.run(training_iterator.string_handle())
-    heldout_handle = sess.run(heldout_iterator.string_handle())
-    for step in range(FLAGS.max_steps):
-      _ = sess.run([train_op, accuracy_update_op],
-                   feed_dict={handle: train_handle})
+  print(" ... Training convolutional neural network")
+  for epoch in range(FLAGS.num_epochs):
+    epoch_accuracy = []
+    for step, (batch_x, batch_y) in enumerate(train_seq):
+      batch_accuracy = model.train_on_batch(
+          batch_x, batch_y)[1]
+      epoch_accuracy.append(batch_accuracy)
 
       if step % 100 == 0:
-        loss_value, accuracy_value = sess.run(
-            [elbo_loss, accuracy], feed_dict={handle: train_handle})
-        print("Step: {:>3d} Loss: {:.3f} Accuracy: {:.3f}".format(
-            step, loss_value, accuracy_value))
+        print('Epoch: {}, Batch index: {}, Accuracy: {:.3f}'.format(
+            epoch, step, np.mean(epoch_accuracy)))
 
       if (step+1) % FLAGS.viz_steps == 0:
         # Compute log prob of heldout set by averaging draws from the model:
         # p(heldout | train) = int_model p(heldout|model) p(model|train)
         #                   ~= 1/n * sum_{i=1}^n p(heldout | model_i)
         # where model_i is a draw from the posterior p(model|train).
-        probs = np.asarray([sess.run((labels_distribution.probs),
-                                     feed_dict={handle: heldout_handle})
+        print(" ... Running monte carlo inference")
+        probs = np.asarray([model.predict_generator(heldout_seq, verbose=1)
                             for _ in range(FLAGS.num_monte_carlo)])
         mean_probs = np.mean(probs, axis=0)
-
-        image_vals, label_vals = sess.run((images, labels),
-                                          feed_dict={handle: heldout_handle})
-        heldout_lp = np.mean(np.log(mean_probs[np.arange(mean_probs.shape[0]),
-                                               label_vals.flatten()]))
-        print(" ... Held-out nats: {:.3f}".format(heldout_lp))
-
-        qm_vals, qs_vals = sess.run((qmeans, qstds))
+        heldout_log_prob = np.mean(np.log(mean_probs))
+        print(" ... Held-out nats: {:.3f}".format(heldout_log_prob))
 
         if HAS_SEABORN:
+          names = [layer.name for layer in model.layers
+                   if 'flipout' in layer.name]
+          qm_vals = [layer.kernel_posterior.mean()
+                     for layer in model.layers
+                     if 'flipout' in layer.name]
+          qs_vals = [layer.kernel_posterior.stddev()
+                     for layer in model.layers
+                     if 'flipout' in layer.name]
           plot_weight_posteriors(names, qm_vals, qs_vals,
                                  fname=os.path.join(
                                      FLAGS.model_dir,
-                                     "step{:05d}_weights.png".format(step)))
-
-          plot_heldout_prediction(image_vals, probs,
+                                     "epoch{}_step{:05d}_weights.png".format(
+                                         epoch, step)))
+          plot_heldout_prediction(heldout_seq.images, probs,
                                   fname=os.path.join(
                                       FLAGS.model_dir,
-                                      "step{:05d}_pred.png".format(step)),
+                                      "epoch{}_step{}_pred.png".format(
+                                          epoch, step)),
                                   title="mean heldout logprob {:.2f}"
-                                  .format(heldout_lp))
+                                  .format(heldout_log_prob))
+
 
 if __name__ == "__main__":
-  tf.compat.v1.app.run()
+  app.run(main)
