--- tensorflow_probability/examples/keras_bayesian_neural_network.py
@@ -1,360 +1,0 @@
-# Copyright 2018 The TensorFlow Probability Authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ============================================================================
-"""Trains a Bayesian neural network to classify MNIST digits.
-
-The architecture is LeNet-5 [1].
-
-#### References
-
-[1]: Yann LeCun, Leon Bottou, Yoshua Bengio, and Patrick Haffner.
-     Gradient-based learning applied to document recognition.
-     _Proceedings of the IEEE_, 1998.
-     http://yann.lecun.com/exdb/publis/pdf/lecun-01a.pdf
-"""
-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-import os
-import warnings
-
-# Dependency imports
-from absl import flags, app
-import matplotlib
-matplotlib.use("Agg")
-from matplotlib import figure  # pylint: disable=g-import-not-at-top
-from matplotlib.backends import backend_agg
-import numpy as np
-import tensorflow as tf
-import tensorflow_probability as tfp
-
-# TODO(b/78137893): Integration tests currently fail with seaborn imports.
-warnings.simplefilter(action="ignore")
-
-try:
-  import seaborn as sns  # pylint: disable=g-import-not-at-top
-  HAS_SEABORN = True
-except ImportError:
-  HAS_SEABORN = False
-
-TFD = tfp.distributions
-
-IMAGE_SHAPE = [28, 28, 1]
-NUM_TRAIN_EXAMPLES = 60000
-NUM_HELDOUT_EXAMPLES = 10000
-NUM_CLASSES = 10
-
-flags.DEFINE_float("learning_rate",
-                   default=0.001,
-                   help="Initial learning rate.")
-flags.DEFINE_integer("num_epochs",
-                     default=10,
-                     help="Number of training steps to run.")
-flags.DEFINE_integer("batch_size",
-                     default=128,
-                     help="Batch size.")
-flags.DEFINE_string("data_dir",
-                    default=os.path.join(os.getenv("TEST_TMPDIR", "/tmp"),
-                                         "bayesian_neural_network/data"),
-                    help="Directory where data is stored (if using real data).")
-flags.DEFINE_string(
-    "model_dir",
-    default=os.path.join(os.getenv("TEST_TMPDIR", "/tmp"),
-                         "bayesian_neural_network/"),
-    help="Directory to put the model's fit.")
-flags.DEFINE_integer("viz_steps",
-                     default=400,
-                     help="Frequency at which save visualizations.")
-flags.DEFINE_integer("num_monte_carlo",
-                     default=50,
-                     help="Network draws to compute predictive probabilities.")
-flags.DEFINE_bool("fake_data",
-                  default=False,
-                  help="If true, uses fake data. Defaults to real data.")
-
-FLAGS = flags.FLAGS
-
-
-def plot_weight_posteriors(names, qm_vals, qs_vals, fname):
-  """Save a PNG plot with histograms of weight means and stddevs.
-
-  Args:
-    names: A Python `iterable` of `str` variable names.
-    qm_vals: A Python `iterable`, the same length as `names`,
-    whose elements are Numpy `array`s, of any shape, containing
-    posterior means of weight varibles.
-    qs_vals: A Python `iterable`, the same length as `names`,
-    whose elements are Numpy `array`s, of any shape, containing
-    posterior standard deviations of weight varibles.
-    fname: Python `str` filename to save the plot to.
-  """
-  fig = figure.Figure(figsize=(6, 3))
-  canvas = backend_agg.FigureCanvasAgg(fig)
-
-  ax = fig.add_subplot(1, 2, 1)
-  for n, qm in zip(names, qm_vals):
-    sns.distplot(qm.numpy().flatten(), ax=ax, label=n)
-  ax.set_title("weight means")
-  ax.set_xlim([-1.5, 1.5])
-  ax.legend()
-
-  ax = fig.add_subplot(1, 2, 2)
-  for n, qs in zip(names, qs_vals):
-    sns.distplot(qs.numpy().flatten(), ax=ax)
-  ax.set_title("weight stddevs")
-  ax.set_xlim([0, 1.])
-
-  fig.tight_layout()
-  canvas.print_figure(fname, format="png")
-  print("saved {}".format(fname))
-
-
-def plot_heldout_prediction(input_vals, probs,
-                            fname, n=10, title=""):
-  """Save a PNG plot visualizing posterior uncertainty on heldout data.
-
-  Args:
-    input_vals: A `float`-like Numpy `array` of shape
-    `[num_heldout] + IMAGE_SHAPE`, containing heldout input images.
-    probs: A `float`-like Numpy array of shape `[num_monte_carlo,
-    num_heldout, num_classes]` containing Monte Carlo samples of
-    class probabilities for each heldout sample.
-    fname: Python `str` filename to save the plot to.
-    n: Python `int` number of datapoints to vizualize.
-    title: Python `str` title for the plot.
-  """
-  fig = figure.Figure(figsize=(9, 3*n))
-  canvas = backend_agg.FigureCanvasAgg(fig)
-  for i in range(n):
-    ax = fig.add_subplot(n, 3, 3*i + 1)
-    ax.imshow(input_vals[i, :].reshape(IMAGE_SHAPE[:-1]), interpolation="None")
-
-    ax = fig.add_subplot(n, 3, 3*i + 2)
-    for prob_sample in probs:
-      sns.barplot(np.arange(10), prob_sample[i, :], alpha=0.1, ax=ax)
-      ax.set_ylim([0, 1])
-    ax.set_title("posterior samples")
-
-    ax = fig.add_subplot(n, 3, 3*i + 3)
-    sns.barplot(np.arange(10), np.mean(probs[:, i, :], axis=0), ax=ax)
-    ax.set_ylim([0, 1])
-    ax.set_title("predictive probs")
-  fig.suptitle(title)
-  fig.tight_layout()
-
-  canvas.print_figure(fname, format="png")
-  print("saved {}".format(fname))
-
-
-def create_model():
-  """
-  Creates a Keras model using the LeNet-5 architecture.
-
-  Returns:
-      model: Compiled Keras model.
-  """
-  # KL divergence weighted by the number of training samples, using
-  # lambda function to pass as input to the kernel_divergence_fn on
-  # flipout layers.
-  kl_divergence_function = lambda q, p, _: TFD.kl_divergence(q, p) / \
-      tf.cast(NUM_TRAIN_EXAMPLES, dtype=tf.float32)
-
-  # Define a LeNet-5 model using three convolutional (with max pooling)
-  # and two fully connected dense layers. We use the Flipout
-  # Monte Carlo estimator for these layers, which enables lower variance
-  # stochastic gradients than naive reparameterization.
-  model = tf.keras.models.Sequential()
-  model.add(tfp.layers.Convolution2DFlipout(
-      6, kernel_size=5, padding="SAME",
-      kernel_divergence_fn=kl_divergence_function,
-      activation=tf.nn.relu))
-  model.add(tf.keras.layers.MaxPooling2D(
-      pool_size=[2, 2], strides=[2, 2],
-      padding="SAME"))
-  model.add(tfp.layers.Convolution2DFlipout(
-      16, kernel_size=5, padding="SAME",
-      kernel_divergence_fn=kl_divergence_function,
-      activation=tf.nn.relu))
-  model.add(tf.keras.layers.MaxPooling2D(
-      pool_size=[2, 2], strides=[2, 2],
-      padding="SAME"))
-  model.add(tfp.layers.Convolution2DFlipout(
-      120, kernel_size=5, padding="SAME",
-      kernel_divergence_fn=kl_divergence_function,
-      activation=tf.nn.relu))
-  model.add(tf.keras.layers.Flatten())
-  model.add(tfp.layers.DenseFlipout(
-      84, kernel_divergence_fn=kl_divergence_function,
-      activation=tf.nn.relu))
-  model.add(tfp.layers.DenseFlipout(
-      NUM_CLASSES, kernel_divergence_fn=kl_divergence_function,
-      activation=tf.nn.softmax))
-
-  # Model compilation.
-  optimizer = tf.keras.optimizers.Adam(lr=FLAGS.learning_rate)
-  # We use the categorical_crossentropy loss since the MNIST dataset contains
-  # ten labels. The Keras API will then automatically add the
-  # Kullback-Leibler divergence (contained on the individual layers of
-  # the model), to the cross entropy loss, effectively
-  # calcuating the Evidence Lower Bound Loss (ELBO)
-  model.compile(optimizer, loss='categorical_crossentropy',
-                metrics=['accuracy'])
-  return model
-
-
-class MNISTSequence(tf.keras.utils.Sequence):
-  """
-  Class inheriting from the Keras Sequence class for producing
-  a sequence of data (to be used with a model.(fit|evaluate|predict)_generator
-  function).
-
-  Args:
-    data: Tuple of numpy `array` instances, the first representing images and
-          the second labels.
-    batch_size: Integer, number of elements in each training batch.
-    fake_data: Boolean, for creating MNIST-like fake data.
-  """
-  def __init__(self, data=None, batch_size=128, size=None):
-    if data:
-      images, labels = data
-    else:
-      images, labels = MNISTSequence.__generate_fake_data(
-          size, NUM_CLASSES)
-    self.images, self.labels = MNISTSequence.__preprocessing(
-        images, labels)
-    self.batch_size = batch_size
-
-  @staticmethod
-  def __generate_fake_data(num_images, num_classes):
-    '''
-    Generate fake data in the shape of the MNIST dataset for unittest.
-    Args:
-      num_images: Integer, the number of fake images to be generated.
-    Returns:
-      images: Numpy `array` representing the fake image data. The
-              shape of the array will be (num_images, 28, 28).
-      labels: Numpy `array` of integers, where each entry will be
-              assigned a unique integer.
-    '''
-    images = np.random.randint(low=0, high=256,
-                               size=(num_images, IMAGE_SHAPE[0],
-                                     IMAGE_SHAPE[1]))
-    labels = np.random.randint(low=0, high=num_classes,
-                               size=num_images)
-    return images, labels
-
-  @staticmethod
-  def __preprocessing(images, labels):
-    '''
-    Preprocessing for image and labels data.
-
-    Args:
-      images: Numpy `array` representing the image data.
-      labels: Numpy `array` representing the labels data (range 0-9).
-
-    Returns:
-      images: Numpy `array` representing the image data, normalized
-              and expanded for convolutional network input.
-      labels: Numpy `array` representing the labels data (range 0-9),
-              as one-hot (categorical) values.
-    '''
-    images = images/255.
-    images = np.expand_dims(images, axis=4)
-
-    labels = tf.keras.utils.to_categorical(labels)
-    return images, labels
-
-  def __len__(self):
-    return int(np.ceil(len(self.images)/self.batch_size))
-
-  def __getitem__(self, idx):
-    batch_x = self.images[idx*self.batch_size:(idx+1)*self.batch_size]
-    batch_y = self.labels[idx*self.batch_size:(idx+1)*self.batch_size]
-    return batch_x, batch_y
-
-
-def main(argv):
-  del argv  # unused
-  if tf.io.gfile.exists(FLAGS.model_dir):
-    tf.compat.v1.logging.warning(
-        "Warning: deleting old log directory at {}".format(FLAGS.model_dir))
-    tf.io.gfile.rmtree(FLAGS.model_dir)
-  tf.io.gfile.makedirs(FLAGS.model_dir)
-
-  if FLAGS.fake_data:
-    train_seq = MNISTSequence(batch_size=FLAGS.batch_size,
-                              size=NUM_TRAIN_EXAMPLES)
-    heldout_seq = MNISTSequence(batch_size=FLAGS.batch_size,
-                                size=NUM_HELDOUT_EXAMPLES)
-  else:
-    train_set, heldout_set = tf.keras.datasets.mnist.load_data()
-    train_seq = MNISTSequence(data=train_set, batch_size=FLAGS.batch_size)
-    heldout_seq = MNISTSequence(data=heldout_set, batch_size=FLAGS.batch_size)
-
-  model = create_model()
-
-  print(" ... Training convolutional neural network")
-  for epoch in range(FLAGS.num_epochs):
-    epoch_accuracy = []
-    for step, (batch_x, batch_y) in enumerate(train_seq):
-      # Eager mode returns a Tensor objest and not a scalar
-      # value for the loss, therefore only the mean accuracy
-      # is displayed.
-      batch_accuracy = model.train_on_batch(
-          batch_x, batch_y)[1]
-      epoch_accuracy.append(batch_accuracy)
-
-      if step % 100 == 0:
-        print('Epoch: {}, Batch index: {}, Accuracy: {:.3f}'.format(
-            epoch, step, np.mean(epoch_accuracy)))
-
-      if (step+1) % FLAGS.viz_steps == 0:
-        # Compute log prob of heldout set by averaging draws from the model:
-        # p(heldout | train) = int_model p(heldout|model) p(model|train)
-        #                   ~= 1/n * sum_{i=1}^n p(heldout | model_i)
-        # where model_i is a draw from the posterior p(model|train).
-        print(" ... Running monte carlo inference")
-        probs = np.asarray([model.predict_generator(heldout_seq, verbose=1)
-                            for _ in range(FLAGS.num_monte_carlo)])
-        mean_probs = np.mean(probs, axis=0)
-        heldout_log_prob = np.mean(np.log(mean_probs))
-        print(" ... Held-out nats: {:.3f}".format(heldout_log_prob))
-
-        if HAS_SEABORN:
-          names = [layer.name for layer in model.layers
-                   if 'flipout' in layer.name]
-          qm_vals = [layer.kernel_posterior.mean()
-                     for layer in model.layers
-                     if 'flipout' in layer.name]
-          qs_vals = [layer.kernel_posterior.stddev()
-                     for layer in model.layers
-                     if 'flipout' in layer.name]
-          plot_weight_posteriors(names, qm_vals, qs_vals,
-                                 fname=os.path.join(
-                                     FLAGS.model_dir,
-                                     "epoch{}_step{:05d}_weights.png".format(
-                                         epoch, step)))
-          plot_heldout_prediction(heldout_seq.images, probs,
-                                  fname=os.path.join(
-                                      FLAGS.model_dir,
-                                      "epoch{}_step{}_pred.png".format(
-                                          epoch, step)),
-                                  title="mean heldout logprob {:.2f}"
-                                  .format(heldout_log_prob))
-
-
-if __name__ == "__main__":
-  app.run(main)
