+++ tensorflow_probability/examples/logistic_regression.py
@@ -21,22 +21,24 @@ from __future__ import print_function
 import os
 
 # Dependency imports
-from absl import flags
+from absl import flags, app
 from matplotlib import cm
 from matplotlib import figure
 from matplotlib.backends import backend_agg
 import numpy as np
-import tensorflow as tf
+import tensorflow.compat.v2 as tf
 import tensorflow_probability as tfp
+
+tf.enable_v2_behavior()
 
 tfd = tfp.distributions
 
 flags.DEFINE_float("learning_rate",
                    default=0.01,
                    help="Initial learning rate.")
-flags.DEFINE_integer("max_steps",
-                     default=1500,
-                     help="Number of training steps to run.")
+flags.DEFINE_integer("num_epochs",
+                     default=50,
+                     help="Number of epochs to run.")
 flags.DEFINE_integer("batch_size",
                      default=32,
                      help="Batch size.")
@@ -53,37 +55,8 @@ flags.DEFINE_integer("num_monte_carlo",
                      help="Monte Carlo samples to visualize weight posterior.")
 
 FLAGS = flags.FLAGS
-
-
-def toy_logistic_data(num_examples, input_size=2, weights_prior_stddev=5.0):
-  """Generates synthetic data for binary classification.
-
-  Args:
-    num_examples: The number of samples to generate (scalar Python `int`).
-    input_size: The input space dimension (scalar Python `int`).
-    weights_prior_stddev: The prior standard deviation of the weight
-      vector. (scalar Python `float`).
-
-  Returns:
-    random_weights: Sampled weights as a Numpy `array` of shape
-      `[input_size]`.
-    random_bias: Sampled bias as a scalar Python `float`.
-    design_matrix: Points sampled uniformly from the cube `[-1,
-       1]^{input_size}`, as a Numpy `array` of shape `(num_examples,
-       input_size)`.
-    labels: Labels sampled from the logistic model `p(label=1) =
-      logistic(dot(features, random_weights) + random_bias)`, as a Numpy
-      `int32` `array` of shape `(num_examples, 1)`.
-  """
-  random_weights = weights_prior_stddev * np.random.randn(input_size)
-  random_bias = np.random.randn()
-  design_matrix = np.random.rand(num_examples, input_size) * 2 - 1
-  logits = np.reshape(
-      np.dot(design_matrix, random_weights) + random_bias,
-      (-1, 1))
-  p_labels = 1. / (1 + np.exp(-logits))
-  labels = np.int32(p_labels > np.random.rand(num_examples, 1))
-  return random_weights, random_bias, np.float32(design_matrix), labels
+# The dimensions of the example data, ie shape=(256, 2)
+NUM_DIMENSIONS = 2
 
 
 def visualize_decision(features, labels, true_w_b, candidate_w_bs, fname):
@@ -92,12 +65,12 @@ def visualize_decision(features, labels, true_w_b, candidate_w_bs, fname):
   Args:
     features: Input points, as a Numpy `array` of shape `[num_examples, 2]`.
     labels: Numpy `float`-like array of shape `[num_examples, 1]` giving a
-      label for each point.
+    label for each point.
     true_w_b: A `tuple` `(w, b)` where `w` is a Numpy array of
-       shape `[2]` and `b` is a scalar `float`, interpreted as a
-       decision rule of the form `dot(features, w) + b > 0`.
+    shape `[2]` and `b` is a scalar `float`, interpreted as a
+    decision rule of the form `dot(features, w) + b > 0`.
     candidate_w_bs: Python `iterable` containing tuples of the same form as
-       true_w_b.
+    true_w_b.
     fname: The filename to save the plot as a PNG image (Python `str`).
   """
   fig = figure.Figure(figsize=(6, 6))
@@ -131,94 +104,130 @@ def visualize_decision(features, labels, true_w_b, candidate_w_bs, fname):
   print("saved {}".format(fname))
 
 
-def build_input_pipeline(x, y, batch_size):
-  """Build a Dataset iterator for supervised classification.
+def toy_logistic_data(num_examples, input_size=2, weights_prior_stddev=5.0):
+  """Generates synthetic data for binary classification.
 
   Args:
-    x: Numpy `array` of features, indexed by the first dimension.
-    y: Numpy `array` of labels, with the same first dimension as `x`.
-    batch_size: Number of elements in each training batch.
+    num_examples: The number of samples to generate (scalar Python `int`).
+    input_size: The input space dimension (scalar Python `int`).
+    weights_prior_stddev: The prior standard deviation of the weight
+        vector. (scalar Python `float`).
 
   Returns:
-    batch_features: `Tensor` feed  features, of shape
-      `[batch_size] + x.shape[1:]`.
-    batch_labels: `Tensor` feed of labels, of shape
-      `[batch_size] + y.shape[1:]`.
+    random_weights: Sampled weights as a Numpy `array` of shape
+        `[input_size]`.
+    random_bias: Sampled bias as a scalar Python `float`.
+    design_matrix: Points sampled uniformly from the cube `[-1,
+        1]^{input_size}`, as a Numpy `array` of shape `(num_examples,
+        input_size)`.
+    labels: Labels sampled from the logistic model `p(label=1) =
+        logistic(dot(features, random_weights) + random_bias)`, as a Numpy
+        `int32` `array` of shape `(num_examples, 1)`.
   """
-  training_dataset = tf.data.Dataset.from_tensor_slices((x, y))
-  training_batches = training_dataset.repeat().batch(batch_size)
-  training_iterator = tf.compat.v1.data.make_one_shot_iterator(training_batches)
-  batch_features, batch_labels = training_iterator.get_next()
-  return batch_features, batch_labels
+  random_weights = weights_prior_stddev * np.random.randn(input_size)
+  random_bias = np.random.randn()
+  design_matrix = np.random.rand(num_examples, input_size) * 2 - 1
+  logits = np.reshape(
+      np.dot(design_matrix, random_weights) + random_bias,
+      (-1, 1))
+  p_labels = 1. / (1 + np.exp(-logits))
+  labels = np.int32(p_labels > np.random.rand(num_examples, 1))
+  return random_weights, random_bias, np.float32(design_matrix), labels
 
 
-def main(argv):
-  del argv  # unused
-  if tf.io.gfile.exists(FLAGS.model_dir):
-    tf.compat.v1.logging.warning(
-        "Warning: deleting old log directory at {}".format(FLAGS.model_dir))
-    tf.io.gfile.rmtree(FLAGS.model_dir)
-  tf.io.gfile.makedirs(FLAGS.model_dir)
+class ToyDataSequence(tf.keras.utils.Sequence):
+  '''
+  Class inheriting from the Keras Sequence class for creating
+  a sequence of data (to be used with a model.fit_generator
+  function).
 
-  # Generate (and visualize) a toy classification dataset.
-  w_true, b_true, x, y = toy_logistic_data(FLAGS.num_examples, 2)
-  features, labels = build_input_pipeline(x, y, FLAGS.batch_size)
+  Args:
+    features: Numpy `array` of features, indexed by the first dimension.
+    labels: Numpy `array` of features, with the same first dimension as
+            `features`.
+    batch_size: Integer, number of elements in each training batch.
+
+  '''
+  def __init__(self, features, labels, batch_size):
+    self.features = features
+    self.labels = labels
+    self.batch_size = batch_size
+
+  def __len__(self):
+    return int(np.ceil(len(self.features)/self.batch_size))
+
+  def __getitem__(self, idx):
+    batch_x = self.features[self.batch_size*idx:self.batch_size*(idx+1)]
+    batch_y = self.labels[self.batch_size*idx:self.batch_size*(idx+1)]
+    return batch_x, batch_y
+
+
+def create_model(num_samples, num_dimensions):
+  '''
+  Creates a Keras model for logistic regression.
+
+  Args:
+   num_dimensions: Integer for number of features in dataset.
+   num_samples: Integer for number of training samples.
+
+  Returns:
+    model: Compiled Keras model.
+  '''
+  # KL divergence weighted by the number of training samples, using
+  # lambda function to pass as input to the kernel_divergence_fn on
+  # flipout layers.
+  kl_divergence_function = lambda q, p, _: tfd.kl_divergence(q, p) / \
+      tf.cast(num_samples, dtype=tf.float32)
 
   # Define a logistic regression model as a Bernoulli distribution
   # parameterized by logits from a single linear layer. We use the Flipout
   # Monte Carlo estimator for the layer: this enables lower variance
   # stochastic gradients than naive reparameterization.
-  with tf.compat.v1.name_scope("logistic_regression", values=[features]):
-    layer = tfp.layers.DenseFlipout(
-        units=1,
-        activation=None,
-        kernel_posterior_fn=tfp.layers.default_mean_field_normal_fn(),
-        bias_posterior_fn=tfp.layers.default_mean_field_normal_fn())
-    logits = layer(features)
-    labels_distribution = tfd.Bernoulli(logits=logits)
+  input_layer = tf.keras.layers.Input(shape=num_dimensions)
+  dense_layer = tfp.layers.DenseFlipout(
+      units=1,
+      activation='sigmoid',
+      kernel_posterior_fn=tfp.layers.default_mean_field_normal_fn(),
+      bias_posterior_fn=tfp.layers.default_mean_field_normal_fn(),
+      kernel_divergence_fn=kl_divergence_function)(input_layer)
 
-  # Compute the -ELBO as the loss, averaged over the batch size.
-  neg_log_likelihood = -tf.reduce_mean(
-      input_tensor=labels_distribution.log_prob(labels))
-  kl = sum(layer.losses) / FLAGS.num_examples
-  elbo_loss = neg_log_likelihood + kl
+  # Model compilation.
+  model = tf.keras.Model(inputs=input_layer, outputs=dense_layer)
+  optimizer = tf.keras.optimizers.Adam(lr=FLAGS.learning_rate)
+  # We use the binary_crossentropy loss since this toy example contains
+  # two labels. The Keras API will then automatically add the
+  # Kullback-Leibler divergence (contained on the individual layers of
+  # the model), to the cross entropy loss, effectively
+  # calcuating the (negated) Evidence Lower Bound Loss (ELBO)
+  model.compile(optimizer, loss='binary_crossentropy', metrics=['accuracy'])
+  return model
 
-  # Build metrics for evaluation. Predictions are formed from a single forward
-  # pass of the probabilistic layers. They are cheap but noisy predictions.
-  predictions = tf.cast(logits > 0, dtype=tf.int32)
-  accuracy, accuracy_update_op = tf.compat.v1.metrics.accuracy(
-      labels=labels, predictions=predictions)
 
-  with tf.compat.v1.name_scope("train"):
-    optimizer = tf.compat.v1.train.AdamOptimizer(
-        learning_rate=FLAGS.learning_rate)
-    train_op = optimizer.minimize(elbo_loss)
+def main(argv):
+  del argv
+  if tf.io.gfile.exists(FLAGS.model_dir):
+    print("Warning: deleting old log directory at {}".format(
+        FLAGS.model_dir))
+    tf.io.gfile.rmtree(FLAGS.model_dir)
+  tf.io.gfile.makedirs(FLAGS.model_dir)
 
-  init_op = tf.group(tf.compat.v1.global_variables_initializer(),
-                     tf.compat.v1.local_variables_initializer())
+  # Generate a toy classification dataset.
+  w_true, b_true, features, labels = toy_logistic_data(FLAGS.num_examples)
+  toy_logistic_sequence = ToyDataSequence(features, labels, FLAGS.batch_size)
 
-  with tf.compat.v1.Session() as sess:
-    sess.run(init_op)
+  # Define and train a bayesian logistic regression model.
+  model = create_model(FLAGS.num_examples, NUM_DIMENSIONS)
+  model.fit_generator(toy_logistic_sequence, epochs=FLAGS.num_epochs,
+                      shuffle=True)
+  # Visualize some draws from the weights posterior.
+  candidate_w_bs = [(model.layers[-1].kernel_posterior.sample().numpy(),
+                     model.layers[-1].bias_posterior.sample().numpy())
+                    for _ in range(FLAGS.num_monte_carlo)]
+  visualize_decision(features, labels, (w_true, b_true),
+                     candidate_w_bs,
+                     fname=os.path.join(FLAGS.model_dir,
+                                        "weights_inferred.png"))
 
-    # Fit the model to data.
-    for step in range(FLAGS.max_steps):
-      _ = sess.run([train_op, accuracy_update_op])
-      if step % 100 == 0:
-        loss_value, accuracy_value = sess.run([elbo_loss, accuracy])
-        print("Step: {:>3d} Loss: {:.3f} Accuracy: {:.3f}".format(
-            step, loss_value, accuracy_value))
-
-    # Visualize some draws from the weights posterior.
-    w_draw = layer.kernel_posterior.sample()
-    b_draw = layer.bias_posterior.sample()
-    candidate_w_bs = []
-    for _ in range(FLAGS.num_monte_carlo):
-      w, b = sess.run((w_draw, b_draw))
-      candidate_w_bs.append((w, b))
-    visualize_decision(x, y, (w_true, b_true),
-                       candidate_w_bs,
-                       fname=os.path.join(FLAGS.model_dir,
-                                          "weights_inferred.png"))
 
 if __name__ == "__main__":
-  tf.compat.v1.app.run()
+  app.run(main)
